+++ lib/utility/bsp/bsp.dart
@@ -7,7 +7,7 @@ import 'package:push_puzzle/utility/bsp/dungeon_config.dart';
 import 'package:push_puzzle/utility/bsp/dungeon_processer.dart';
 import 'package:push_puzzle/utility/bsp/partition.dart';
 import 'package:push_puzzle/utility/bsp/room_creator_visitor.dart';
-import 'package:push_puzzle/utility/bsp/util.dart';
+import 'package:push_puzzle/utility/bsp/Tracer.dart';
 
 
 Future<void> main() async {
