+++ lib/utility/bsp/dungeon_processer.dart
@@ -3,11 +3,11 @@ import 'package:push_puzzle/utility/bsp/dungeon_config.dart';
 import 'package:push_puzzle/utility/bsp/partition.dart';
 import 'package:push_puzzle/utility/bsp/partition_visitor.dart';
 import 'package:push_puzzle/utility/bsp/room_creator_visitor.dart';
-import 'package:push_puzzle/utility/bsp/util.dart';
+import 'package:push_puzzle/utility/bsp/Tracer.dart';
 
 class DungeonProcessor {
   DungeonConfig d = DungeonConfig();
-  Util u = Util();
+  CacheTracer u = CacheTracer();
   RoomCreatorVisitor roomCreator = RoomCreatorVisitor();
   late Partition root;
   late List<List<int>> initialRect;
@@ -19,8 +19,6 @@ class DungeonProcessor {
         depth: d.initialDepth, isRoot: d.initialIsRoot, rect: initialRect, name: d.rootName);
     PartitionVisitor visitor = PartitionVisitor();
     visitor.visit(root);
-
-    root.traceInfo();
 
     u.trace2d(root.getMergedRect());
